[REFACTORED]: <ProductQuantitySelector> replaces <NumberSelector>. Translated all the update quantity logic to the <ProductQuantitySelector>.
